+++ app/main/handlers/chaos_maker_api.js
@@ -1,8 +1,5 @@
 const {ipcMain} = require("electron")
-const {htmlTemplateDir} = require("../filePath")
-const compressing = require("compressing")
-const fs = require("fs")
-const path = require("path")
+
 const handlerHelper = require("./handleStreamWithContext");
