+++ app/main/handlers/http_flow_api.js
@@ -1,9 +1,4 @@
 const {ipcMain} = require("electron")
-const {htmlTemplateDir} = require("../filePath")
-const compressing = require("compressing")
-const fs = require("fs")
-const path = require("path")
-const handlerHelper = require("./handleStreamWithContext");
 
 
 module.exports = (win, getClient) => {
