+++ app/main/handlers/port_scan_api.js
@@ -1,7 +1,6 @@
 const {ipcMain} = require("electron")
 const handlerHelper = require("./handleStreamWithContext");
-const xlsx = require("node-xlsx");
-const FS = require("fs");
+
 
 
 module.exports = (win, getClient) => {
@@ -91,39 +90,4 @@ module.exports = (win, getClient) => {
         handlerHelper.registerHandler(win, stream, streamRecoverSimpleDetectUnfinishedTaskMap, token)
     })
 
-
-    // electron
-    const asyncFetchFileContent = (params) => {
-        return new Promise((resolve, reject) => {
-            const type = params.split(".").pop()
-            const typeArr = ['csv', 'xls', 'xlsx']
-            // 读取Excel
-            if (typeArr.includes(type)) {
-                // 读取xlsx
-                try {
-                    const obj = xlsx.parse(params)
-                    resolve(obj)
-                } catch (error) {
-                    reject(err)
-                }
-            }
-            // 读取txt
-            else {
-                FS.readFile(params, 'utf-8', function (err, data) {
-                    if (err) {
-                        reject(err)
-                    } else {
-                        resolve(data)
-                    }
-                });
-            }
-
-        })
-    }
-
-
-    // 获取URL的IP地址
-    ipcMain.handle("fetch-file-content", async (e, params) => {
-        return await asyncFetchFileContent(params)
-    })
 }
